# frontend/android/openapi.yml
openapi: 3.0.0
servers:
  - url: https://api.chimege.com/v1.2
    description: Сервер
info:
  description: |
    # Танилцуулга
    Энэхүү зааварт Чимэгэ Системс ХХК-с гаргасан бичвэрээс яриа руу хөрвүүлэх, ярианаас бичвэр рүү хөрвүүлэх
    болон үүнтэй холбоотой туслах API-үүд болон Болорсофт ХХК-с гаргасан бичвэрийн алдаа шалгах, үг санал болгох болон кимо хөрвүүлэгч API-үүдийг
    ашиглах зааврууд багтсан болно.

    # Нэвтрэлт
    Чимэгэ API ашиглахын тулд танд идэвхтэй токен хэрэгтэй бөгөөд үүнийг та бидэнтэй холбоо барин үйлчилгээний 
    эрхээ авсны дараа [console.chimege.com](https://console.chimege.com) хаягаар орон үүсгэж авах юм.

    <SecurityDefinitions />

  version: 1.2.0
  title: Чимэгэ API
  contact:
    name: Холбоо барих
    email: "info@chimege.mn"
    url: https://chimege.mn
  x-logo:
    url: '/v1.2/logo.png'
    altText: Chimege Systems LLC
...
paths:
  /transcribe:
    post: 
      parameters:
        - name: token
          in: header
          description: "Таны API ашиглах токен"
          required: false
          schema:
            type: string
        - name: punctuate
          in: header
          description: "Хөрвөсөн бичвэрт цэг, таслал тавьж форматлах эсэх"
          required: false
          schema:
            type: boolean
            default: "false"
      tags:
        - transcribe
      summary: Яриаг бичвэр рүү хөрвүүлэх
      description: Яриаг бичвэр рүү хөрвүүлнэ. Хөрвөсөн бичвэр зөвхөн кирилл жижиг үсэг болон зай агуулсан байна. `Punctuate` параметр ашиглан бичвэрийг форматлах боломжтой. 
      operationId: transcribe
      requestBody:
        description: Хөрвүүлэх аудио
        content:
          application/octet-stream: 
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Амжилттай
          content:
            text/plain: 
              schema:
                type: string
                description: Хөрвөсөн бичвэр
        '400':
          description: Оролтын аудиотой холбоотой алдаа
          headers:
            Error-Code: 
              schema:
                type: integer
              description: |
                Алдааны код.
                * 2000 - Error at receiving audio data. Please check your connection speed and audio data.
                * 2001 - Audio file is too large. Allowed maximum file size is 3MB. Please sample down your audio. 
                * 2002 - Audio file is too small. Minimum file size is 50KB for wav, 2KB for others.
                * 2003 - Audio is too short. Minimum duration is 0.5s.
                * 2004 - Invalid audio encoding. Please check your audio data encoding. Valid format is WAV.
                * 2005 - Failed to convert audio to WAV. Please check your audio data.
        '403':
          description: Токентой холбоотой алдаа
          headers:
            Error-Code: 
              schema:
                type: integer
              description: |
                Алдааны код.
                * 1000 - Invalid API token.
                * 1001 - API token is missing.
                * 1002 - Inactive API token.
                * 1003 - Suspended API token.
        '500':
          description: Таамаглаагүй алдаа гарсан
        '503':
          description: Серверийн ачаалал хэтэрсэн
      x-codeSamples:
        - lang: 'Python'
          source: |
            import requests

            def transcribe(filename):
                with open(filename, 'rb') as f:
                    audio = f.read()

                r = requests.post("https://api.chimege.com/v1.2/transcribe", data=audio, headers={
                    'Content-Type': 'application/octet-stream',
                    'Punctuate': 'true',
                    'Token': 'API_TOKEN',
                })

                return r.content.decode("utf-8")

            print(transcribe('input.wav'))
        - lang: 'Go'
          source: |
            package main

            import (
              "bytes"
              "io/ioutil"
              "log"
              "net/http"
            )

            func main() {
              data, _ := ioutil.ReadFile("input.wav")

              req, _ := http.NewRequest(
                "POST", 
                "https://api.chimege.com/v1.2/transcribe", 
                bytes.NewBuffer(data))

              req.Header.Set("Content-Type", "application/octet-stream")
              req.Header.Set("punctuate", "true")
              req.Header.Set("token", "API_TOKEN")

              client := &http.Client{}

              resp, _ := client.Do(req)
              if err != nil {
                log.Fatal(err)
              }
              defer resp.Body.Close()

              body, _ := ioutil.ReadAll(resp.Body)

              transcription := string(body)
              log.Println(transcription)
            }
...